Add is:inline to PostLayout Twitter script to supress processing hint
src/layouts/PostLayout.astro:99:11 - warning astro(4000): This script will be treated as if it has the `is:inline` directive because it contains an attribute. Therefore, features that require processing (e.g. using TypeScript or npm packages in the script) are unavailable.

See docs for more details: https://docs.astro.build/en/guides/client-side-scripts/#script-processing.

Add the `is:inline` directive explicitly to silence this hint.

99   <script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>
             ~~~~~